requirements.txt
asgiref==3.4.1
boto3==1.20.8
botocore==1.23.8
Django==3.2.8
django-cors-headers==3.10.0
django-storages==1.12.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
gunicorn==23.0.0
jmespath==0.10.0
Pillow==8.4.0
psycopg2==2.9.2
PyJWT==2.3.0
python-dateutil==2.8.2
pytz==2021.3
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
sqlparse==0.5.0
urllib3==2.5.0
whitenoise==5.3.0
